Rename form validation file
